safari-600.7-branch/LayoutTests/svg/filters/feComponentTransfer-style-crash-expected.txt

ALERT: Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1
PASS
